providers/mod.rs: move trait noops into MetadataProvider
Generally in all the providers there are trait(s) that are not
implemented or used by the provider; they all had various methods of
noop'ing. This change unifies the noop handling by implementing the
required fn in the trait definition itself.
