+++ src/providers/cloudstack/network.rs
@@ -4,10 +4,8 @@ use std::collections::HashMap;
 use std::net::IpAddr;
 
 use openssh_keys::PublicKey;
-use slog_scope::warn;
 
 use crate::errors::*;
-use crate::network;
 use crate::providers::MetadataProvider;
 use crate::retry;
 use crate::util;
@@ -93,19 +91,5 @@ impl MetadataProvider for CloudstackNetwork {
         } else {
             Ok(vec![])
         }
-    }
-
-    fn networks(&self) -> Result<Vec<network::Interface>> {
-        Ok(vec![])
-    }
-
-    fn virtual_network_devices(&self) -> Result<Vec<network::VirtualNetDev>> {
-        warn!("virtual network devices metadata requested, but not supported on this platform");
-        Ok(vec![])
-    }
-
-    fn boot_checkin(&self) -> Result<()> {
-        warn!("boot check-in requested, but not supported on this platform");
-        Ok(())
     }
 }
