+++ src/providers/openstack/network.rs
@@ -3,10 +3,8 @@
 use std::collections::HashMap;
 
 use openssh_keys::PublicKey;
-use slog_scope::warn;
 
 use crate::errors::*;
-use crate::network;
 use crate::providers::MetadataProvider;
 use crate::retry;
 
@@ -102,19 +100,5 @@ impl MetadataProvider for OpenstackProviderNetwork {
         }
 
         Ok(out)
-    }
-
-    fn networks(&self) -> Result<Vec<network::Interface>> {
-        Ok(vec![])
-    }
-
-    fn virtual_network_devices(&self) -> Result<Vec<network::VirtualNetDev>> {
-        warn!("virtual network devices metadata requested, but not supported on this platform");
-        Ok(vec![])
-    }
-
-    fn boot_checkin(&self) -> Result<()> {
-        warn!("boot check-in requested, but not supported on this platform");
-        Ok(())
     }
 }
